% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_functions.R
\name{add_entrez_ids_to_esets}
\alias{add_entrez_ids_to_esets}
\title{Function to append entrez-ids to a list of esets. The function uses refseq /
genbank identifies that are present in the esets}
\usage{
add_entrez_ids_to_esets(esets, entrezgene.db)
}
\arguments{
\item{esets}{An ExpressionSet or a list of ExpressionSets.
Assumed to contain a genbank/refseq column.}

\item{entrezgene.db}{An AnnotationDbi object (passed to
multisymbol_to_entrez...)}
}
\description{
Function to append entrez-ids to a list of esets. The function uses refseq /
genbank identifies that are present in the esets
}
