% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{add_gpl_to_eset}
\alias{add_gpl_to_eset}
\title{add_gpl_to_eset}
\usage{
add_gpl_to_eset(eset, gpl_data)
}
\arguments{
\item{eset}{An \code{ExpressionSet}}

\item{gpl_data}{A \code{GPL} from \code{GEOquery} containing
annotation data for the \code{ExpressionSet} \code{eset}}
}
\description{
Replaces the featureData slot of an ExpressionSet with the corresponding
rows of a "GEOquery::GPL" object.
}
