% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_functions.R
\name{build_eset_normaliser}
\alias{build_eset_normaliser}
\title{create a function that can normalise an ExpressionSet (optionally converting
the input dataset to an ExpressionSet beforehand)}
\usage{
build_eset_normaliser(eset_making_fn = identity,
  normalise_fn = c("none", "quantile"),
  log_checking_fn = check_if_pretransformed_eset)
}
\arguments{
\item{eset_making_fn}{A function to convert a dataset to an
ExpressionSet. If the user is already working with an ExpressionSet, just
use \code{identity} (the default). Otherwise, you might use
\code{oligo::rma}.}

\item{normalise_fn}{A function to normalise an ExpressionSet. Options are
`quantile` (for quantile normalisation) and `none` (for do not modify; the
default).}

\item{log_checking_fn}{A function that can determine whether an
ExpressionSet is already log-transformed. This helps prevent you from
log-transforming an already log-transformed dataset, and from forgetting
to log-transform and non-transformed dataset.}
}
\description{
create a function that can normalise an ExpressionSet (optionally converting
the input dataset to an ExpressionSet beforehand)
}
