% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_functions.R
\name{check_if_pretransformed_eset}
\alias{check_if_pretransformed_eset}
\title{Function that performs a few tests to check if a given \code{ExpressionSet}
has already been log-transformed}
\usage{
check_if_pretransformed_eset(eset = NULL,
  range_limit_if_transformed = 1000, skew_threshold = 3)
}
\arguments{
\item{eset}{an \code{ExpressionSet}.}

\item{range_limit_if_transformed}{the largest possible range
between high and low values if the \code{ESet} is already transformed.}

\item{skew_threshold}{if the skewness of all columns of the
dataset is greater than this threshold, the dataset is assumed
untransformed.}
}
\description{
Function that performs a few tests to check if a given \code{ExpressionSet}
has already been log-transformed
}
