% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions.R
\name{contrast_builder}
\alias{contrast_builder}
\title{Builds a function that makes the contrast matrix for a dataset}
\usage{
contrast_builder(x)
}
\arguments{
\item{x}{Either a closure or a list of string-based contrast
definitions. If a closure, this should accept a design matrix and return
a contrast matrix.}
}
\value{
A function that accepts a design matrix / data.frame and
  constructs a contrasts matrix from it.
}
\description{
Note that for a function-as-input, all this does is wrap that function and
check the dimensions and type of the input design against the output
contrast matrix.
}
