% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_functions.R
\name{design_builder}
\alias{design_builder}
\title{Builds a function that makes the design matrix for a dataset}
\usage{
design_builder(treatment_cols = NULL, design_fn = NULL)
}
\arguments{
\item{treatment_cols}{Which columns of the pData should be kept?}

\item{design_fn}{Function for making design matrices.}
}
\description{
User passes in the column labels of the phenoData that are to be used in
  making the design
And passes in a design-function that converts these columns into the design
The functionBuilder checks for the validity of the input (treatment_cols /
design_fn) and the built function checks the validity of the ExpressionSet
that gets passed in to it.
}
