% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_classes.R
\name{DiffexConfig}
\alias{DiffexConfig}
\title{Constructor function for DiffexConfig class}
\usage{
DiffexConfig(design, contrast)
}
\arguments{
\item{design}{Either a function or a design matrix. If it is a function,
then design(ExpressionSet) should be a design matrix.}

\item{contrast}{Either a function or a contrast matrix. If it is a
function the contrast(design_matrix) should be a contrast matrix.}
}
\description{
Constructor function for DiffexConfig class
}
