% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{dl_aryx_raw}
\alias{dl_aryx_raw}
\title{Download raw data from ArrayExpress}
\usage{
dl_aryx_raw(acc, dest_dir = tempdir())
}
\arguments{
\item{acc}{An Array-Express accession number, eg,
"E-MTAB-2893"}

\item{dest_dir}{The parent directory in which all ArrayExpress
datasets are downloaded to. Note that the datasets for \code{acc} all get
stored into \code{`<dest_dir>/<acc>`}}
}
\description{
Download raw data from ArrayExpress
}
