% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{dl_gds_to_gpl}
\alias{dl_gds_to_gpl}
\title{Obtains the GPL accession(s) for a set of input GDS uids from NCBI-eUtils}
\usage{
dl_gds_to_gpl(ids,
  eutils_prefix = paste0("https://eutils.ncbi.nlm.nih.gov/",
  "entrez/eutils/esummary.fcgi?db=gds"))
}
\arguments{
\item{ids}{ABC}

\item{eutils_prefix}{DEF}
}
\description{
Uses NCBI-eUtils. Takes GEO dataset UIDs (of the format 200012345) and
creates a mapping from these UIDs to the (default) GEO platform accession
number used by GEO to annotate the GEO dataset. GPL accessions will be of
the form "GPL1234" and GDS uids will be returned in the same for as input
(ie, 200012345). Does not make an effort to ensure input order is the same
as output order.
}
