% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{dl_geo_processed}
\alias{dl_geo_processed}
\title{Download processed data from GEO for a GSE accession number}
\usage{
dl_geo_processed(acc, dest_dir = file.path("data", "ext", "GEOquery"))
}
\arguments{
\item{acc}{A single GSE accession number of the form
"GSE12345".}

\item{dest_dir}{An existing directory. The function dies if this
directory is missing. Data related to gse-id are stored either in this
directory (the processed data matrix and the annotation data) or in a
subdirectory <gse.id> of this directory (for the raw data).}
}
\value{
A list giving the directories and filenames for all downloaded
  files. The list contains some entries that are only relevant to
  array-express downloads, rather than GEO downloads.
}
\description{
Given a GSE accession number (GSE12345), this function downloads the
processed dataset from the Gene Expression Omnibus as a matrix. Processed
data are stored as \code{dest_dir/<gse_acc>_series_matrix.txt.gz} and any
annotation data is stored in \code{dest_dir/}.
}
