% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{dl_geo_raw}
\alias{dl_geo_raw}
\title{Download raw data from GEO for a GSE accession number}
\usage{
dl_geo_raw(acc, dest_dir = file.path("data", "ext", "GEOquery"))
}
\arguments{
\item{acc}{A single GSE accession number of the form
"GSE12345".}

\item{dest_dir}{An existing directory. The function dies if this
directory is missing. Data related to gse-id are stored either in this
directory (the processed data matrix and the annotation data) or in a
subdirectory <gse.id> of this directory (for the raw data).}
}
\value{
A list giving the directories and filenames for all downloaded
  files. The list contains some names that are relevant to array-express
  downloads, rather than GEO downloads.
}
\description{
Given a GSE identifier, this function downloads the raw data for that
dataset from the Gene Expression Omnibus. Raw data is stored in
<dest_dir>/<gse_acc>/*_RAW.tar.
}
\details{
The function checks whether the raw data has already been
downloaded before pulling it from GEO again.
}
