% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{dl_gpl_annotations}
\alias{dl_gpl_annotations}
\title{Download GPL annotations from GEO}
\usage{
dl_gpl_annotations(gpl_acc, dest_dir = file.path("data", "ext",
  "GEOquery"), annot_gpl = TRUE)
}
\arguments{
\item{gpl_acc}{String. A single GPL accession number, as used
at Gene-Expression Omnibus. Of the form GPL12345.}

\item{dest_dir}{Directory in which the GPL annotation dataset
should be saved. Default is \code{"./data/ext/GEOquery"} and this
directory should exist before calling this function.}

\item{annot_gpl}{Logical. Should the function return the URL for
the most recent (NCBI-constructed) annotations for the platform, or the
user/vendor-submitted 'soft' file.}
}
\description{
TODO: docs.
Note that invalid GPL id kills the function.
Note that invalid combination of annot.gpl and gpl.acc kill the function.
Wrapper around getGEO for GPLs that checks that an GPL dataset is available
  before attempting to download it.
}
