% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_from_env}
\alias{.get_from_env}
\title{Obtain an object from a specific package based on a string}
\usage{
.get_from_env(x)
}
\arguments{
\item{x}{a string in the form "some_object_name" or
"some_package::some_object_name".}
}
\description{
Suppose you provide x = "abc::def", then this will return the `def` function
or variable from the package `abc`. If x = "ghi", then the `ghi` function
or variable from the enclosing environment will be returned
}
