% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{.make_gpl_filename}
\alias{.make_gpl_filename}
\title{Constructs the filename for a GPL dataset}
\usage{
.make_gpl_filename(gpl_acc, annot_gpl = TRUE, for_url = FALSE)
}
\arguments{
\item{gpl_acc}{String. A single GPL accession number, as used
at Gene-Expression Omnibus. Of the form GPL12345.}

\item{annot_gpl}{Logical. Should the function return the URL for
the most recent (NCBI-constructed) annotations for the platform, or the
user/vendor-submitted 'soft' file.}

\item{for_url}{Logical. Is the required filename for use when
searching the NCBI ftp site, or when storing the data locally?}
}
\description{
Constructs the filename for a GPL dataset
}
