% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\name{.make_gpl_url}
\alias{.make_gpl_url}
\title{Converts a GPL accession number into the FTP URL for that annotation
dataset.}
\usage{
.make_gpl_url(gpl_acc, annot_gpl = TRUE,
  ftp_prefix = paste0("ftp://ftp.ncbi.nlm.nih.gov/", "geo/platforms"))
}
\arguments{
\item{gpl_acc}{String. A single GPL accession number, as used
at Gene-Expression Omnibus. Of the form GPL12345.}

\item{annot_gpl}{Logical. Should the function return the URL for
the most recent (NCBI-constructed) annotations for the platform, or the
user/vendor-submitted 'soft' file.}

\item{ftp_prefix}{Prefix for the URL.}
}
\description{
Converts a GPL accession number into the FTP URL for that annotation
dataset.
}
