% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_methods.R
\name{download_gpl_annotations}
\alias{download_gpl_annotations}
\title{Download GPL-annotations from GEO for a MicroarrayDownloadConfig object.}
\usage{
download_gpl_annotations(object)
}
\arguments{
\item{object}{A MicroarrayDownloadConfig object.}
}
\description{
Download GPL-annotations from GEO for a MicroarrayDownloadConfig object.
}
