% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_classes.R
\name{eset_limma_dataset-class}
\alias{eset_limma_dataset-class}
\alias{eset_limma_dataset}
\title{Class to hold ExpressionSets and limma model results}
\usage{
eset_limma_dataset(...)
}
\arguments{
\item{...}{arguments passed to the constructor for
`eset_limma_dataset`.}
}
\description{
Class to hold ExpressionSets and limma model results

wrapper function eset_limma_dataset
}
