% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_classes.R
\name{[}
\alias{[}
\alias{[,eset_limma_dataset,numeric,missing,missing-method}
\title{Row-subsetting operator for `eset_limma_dataset` class}
\arguments{
\item{x}{An `eset_limma_dataset` object}

\item{i}{A subset of the rows in `x`}

\item{j, drop, ...}{Not used}
}
\description{
Row-subsetting operator for `eset_limma_dataset` class
}
