% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_functions.R
\name{filter_and_transform_eset}
\alias{filter_and_transform_eset}
\title{Function that filters and transforms the entries in the \code{exprs} entry
of an \code{ExpressionSet}}
\usage{
filter_and_transform_eset(eset = NULL, log2_transform = TRUE,
  drop_row_if_duplicated = TRUE, drop_row_if_zero_variance = TRUE,
  convert_inf_to_na = TRUE, normalise_method = c("quantile", "median",
  "median_iqr", "none"), drop_row_na_inf_threshold = 0.25)
}
\arguments{
\item{eset}{An \code{ExpressionSet}.}

\item{log2_transform}{BOOLEAN : Should the \code{ExpressionSet} be
log2-transformed?}

\item{drop_row_if_duplicated}{BOOLEAN : Should rows be dropped if
they are (numerical) duplicates of another row?}

\item{drop_row_if_zero_variance}{BOOLEAN : Should rows be dropped
if they are constant?}

\item{convert_inf_to_na}{BOOLEAN : Should infinite vals be
converted to NA?}

\item{normalise_method}{String from "quantile" (quantile-normalise
the dataset, the default), "median" (subtract the medians off),
"median_iqr" (subtract off the medians and divide by the iqr) or "none"
(leave the dataset untouched).}

\item{drop_row_na_inf_threshold}{What fraction of missing (NA) or
infinite values are permitted in a row before that row is dropped from
further analysis.}
}
\description{
Function that filters and transforms the entries in the \code{exprs} entry
of an \code{ExpressionSet}
}
