% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{geo_id_to_accession}
\alias{geo_id_to_accession}
\title{geo_id_to_accession}
\usage{
geo_id_to_accession(geo_ids)
}
\arguments{
\item{geo_ids}{A vector of numerical NCBI::GEO identifiers. May
be given as a numeric vector or a character-vector.}
}
\value{
A vector of NCBI::GEO accession numbers (eg, GPL12345,
  GSE98765, GSM24681). NA is returned for any input identifier that is not
  of the 9-digit 1-, 2- or 3-leading format used for uids by NCBI::GEO.
}
\description{
Converts the numeric 9-mer UIDs used by the NCBI::GEO database into the
GSExxxxx / GPLxxxxxx / GSMxxxxx accession numbers that are also used by that
database.
}
