% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manipulation.R
\name{get_or_else}
\alias{get_or_else}
\title{get_or_else}
\usage{
get_or_else(.x, .field, .default_val, .default_fn)
}
\arguments{
\item{.x}{A list}

\item{.field}{A string. This may or may not be one of the
names in the list .x; if it is the associated value is returned. If not
either .default_val or .default_fn(.field) is returned.}

\item{.default_val}{A value that is to be returned if .field is not
in .x. Only one of .default_val and .default_fn may be defined.}

\item{.default_fn}{A single argument function that is evaluated if
.field is not in .x. The value .field is passed into this function as an
argument.  Only one of .default_val and .default_fn may be defined.}
}
\description{
Given a list .x, this function looks to see if the field .field is defined
within that list. If it is defined, the value for that entry is returned. If
.field is not defined, then the function either returns the value
.default_val or the value .default_fn(.field). The latter is lazily
evaluated, ie, it's only computed if .field is not in .x.
}
