% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_functions.R
\name{get_refseq_colnames}
\alias{get_refseq_colnames}
\title{Not exported: checks if a valid refseq column is present in an
  \code{ExpressionSet} and returns a vector of all refseq column names if it
  does; otherwise returns an empty vector}
\usage{
get_refseq_colnames(gset = NULL,
  refseq_col_types = c("RefSeq Transcript ID", "RefSeq.Transcript.ID",
  "GB_LIST", "GB_ACC"))
}
\arguments{
\item{gset}{An \code{ExpressionSet}.}

\item{refseq_col_types}{The choice of refseq column names.}
}
\description{
Not exported: checks if a valid refseq column is present in an
  \code{ExpressionSet} and returns a vector of all refseq column names if it
  does; otherwise returns an empty vector
}
