% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gld_fnBuilder_exptDesign}
\alias{gld_fnBuilder_exptDesign}
\title{Builds a function that makes the design matrix for a dataset}
\usage{
gld_fnBuilder_exptDesign(treatment.cols = NULL, design.fn = NULL)
}
\arguments{
\item{treatment.cols}{Which columns of the pData should be kept?}

\item{design.fn}{Function for making design matrices.}
}
\description{
User passes in the column labels of the phenoData that are to be used in
  making the design
And passes in a design.function that converts these columns into the design
The functionBuilder checks for the validity of the input (treatment.cols /
  design.function) and the built function checks the validity of the
  ExpressionSet / geo-limma-dataset that was passed in
}
