% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gld_fnBuilder_keepSample}
\alias{gld_fnBuilder_keepSample}
\title{Builds a function that returns the col indices of the desired samples}
\usage{
gld_fnBuilder_keepSample(column.filter.fn)
}
\arguments{
\item{column.filter.fn}{Function that decides which samples to keep}
}
\description{
Creates a function that can be applied to a geo-limma-dataset
  or to an ExpressionSet, and returns the column indices of those
  samples that are to kept
This just reduces the boilerplate involved in checking g-l-d or
  ExpressionSet nature of the input
}
