% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gld_fnDefault_exptDesign}
\alias{gld_fnDefault_exptDesign}
\title{Default function for setting up an expt design from the pData of the
  ExpressionSet}
\usage{
gld_fnDefault_exptDesign(geo.limma.dataset = NULL, gset = NULL)
}
\arguments{
\item{geo.limma.dataset}{A geo_limma_dataset.}

\item{gset}{An ExpressionSet.}
}
\description{
Just takes the "title" column of the pData and uses it as a factor.
}
