% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_classes.R
\name{gld_fnDefault_keepProbe}
\alias{gld_fnDefault_keepProbe}
\title{Default function for choosing which probes to keep in an ESet / g-l-d}
\usage{
gld_fnDefault_keepProbe(geo.limma.dataset = NULL, gset = NULL)
}
\arguments{
\item{geo.limma.dataset}{A geo_limma_dataset}

\item{gset}{An ExpressionSet}
}
\description{
Default function for choosing which probes to keep in an ESet / g-l-d
}
