% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gld_fnDefault_keepSample}
\alias{gld_fnDefault_keepSample}
\title{Default function for choosing which samples to keep in a ESet / g-l-d}
\usage{
gld_fnDefault_keepSample(geo.limma.dataset = NULL, gset = NULL)
}
\arguments{
\item{geo.limma.dataset}{An eset_limma_dataset.}

\item{gset}{An ExpressionSet - used preferentially
over ELD@eset.}
}
\description{
Default function for choosing which samples to keep in a ESet / g-l-d
}
