% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_classes.R
\name{gld_fnDefault_transformAndProbeFilter}
\alias{gld_fnDefault_transformAndProbeFilter}
\title{Default function for transforming (etc) a microarray dataset}
\usage{
gld_fnDefault_transformAndProbeFilter(geo.limma.dataset = NULL,
  gset = NULL)
}
\arguments{
\item{geo.limma.dataset}{A geo_limma_dataset}

\item{gset}{An ExpressionSet - used preferentially to
GLD@eset.}
}
\value{
An ExpressionSet
}
\description{
Not exported
}
