% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_functions.R
\name{has_refseq_column}
\alias{has_refseq_column}
\title{Not exported: checks if an ExpressionSet has a valid RefSeq column}
\usage{
has_refseq_column(gset = NULL)
}
\arguments{
\item{gset}{An \code{ExpressionSet}.}
}
\description{
Not exported: checks if an ExpressionSet has a valid RefSeq column
}
