% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_geo_affy_raw}
\alias{import_geo_affy_raw}
\title{import_geo_affy_raw}
\usage{
import_geo_affy_raw(mic, tar = "tar")
}
\arguments{
\item{mic}{A MicroarrayImportConfig object}

\item{tar}{Shell command for tar/untar-ing files. Default
is 'tar' since I can't use absolute paths to /bin/tar.}
}
\description{
import_geo_affy_raw
}
