% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_geo_processed}
\alias{import_geo_processed}
\title{import_geo_processed}
\usage{
import_geo_processed(mic)
}
\arguments{
\item{mic}{A MicroarrayImportConfig object}
}
\description{
import_geo_processed
}
