% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_gpl}
\alias{import_gpl}
\title{import_gpl}
\usage{
import_gpl(gpl_file, gpl_dir)
}
\arguments{
\item{gpl_file}{Filename for a GPL dataset (either a *.soft or a
annot.gz file).}

\item{gpl_dir}{Dirname corresponding to the GPL dataset in
\code{gpl_file}.}
}
\description{
Reads in the microarray annotation data stored in a local copy of a GPL
dataset (as downloaded from NCBI::GEO).
}
