% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_manipulation.R
\name{inf_omit}
\alias{inf_omit}
\title{Function to remove any infinite values from a vector}
\usage{
inf_omit(x)
}
\arguments{
\item{x}{Some vector - only makes sense if \code{x} is
\code{numeric}.}
}
\description{
The returned vector will be shorter than the input vector if any infinite
values are found.
}
