% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{keep_all_entrez_probes}
\alias{keep_all_entrez_probes}
\title{keep_all_entrez_probes}
\usage{
keep_all_entrez_probes(gset)
}
\arguments{
\item{gset}{An \code{ExpressionSet}.}
}
\description{
keep_all_entrez_probes
}
