% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{keep_all_probes}
\alias{keep_all_probes}
\title{keep_all_probes: return the index for each probe on an ExpressionSet}
\usage{
keep_all_probes(gset)
}
\arguments{
\item{gset}{An \code{ExpressionSet}.}
}
\description{
keep_all_probes: return the index for each probe on an ExpressionSet
}
