% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{keep_probe_fn}
\alias{keep_probe_fn}
\title{Function for filtering on microarray probes}
\usage{
keep_probe_fn(geo.limma.dataset = NULL, gset = NULL)
}
\arguments{
\item{geo.limma.dataset}{An eset.limma.dataset}

\item{gset}{An ExpressionSet (overrides use of
geo.limma.dataset)}
}
\description{
Function for filtering on microarray probes
}
