% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{limma_workflow}
\alias{limma_workflow}
\title{limma workflow: .....}
\usage{
limma_workflow(gset = NULL, design_fn = NULL, contrast_fn = NULL)
}
\arguments{
\item{gset}{An ExpressionSet}

\item{design_fn}{A function that creates a design matrix based on
gset as input.}

\item{contrast_fn}{A function that creates a contrasts matrix
based on an input design matrix.}
}
\value{
A list(design, contrast, fits_init, fits).
}
\description{
limma workflow: .....
}
