% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_classes.R
\name{meta-extdata_download_classes-NULL-function}
\alias{meta-extdata_download_classes-NULL-function}
\title{meta-extdata_download_classes-NULL-function}
\description{
Unnecessary description for
  meta-extdata_download_classes-NULL
}
