% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_classes.R
\name{MicroarrayImportConfig}
\alias{MicroarrayImportConfig}
\title{Sets up a class for storing the options for importing datasets from GEO
and/or ArrayExpress}
\usage{
MicroarrayImportConfig(acc, import_method, normalise_method,
  raw_dir = as.character(NA), raw_files = as.character(NA),
  raw_archive = as.character(NA), processed_dir = as.character(NA),
  processed_files = as.character(NA),
  processed_archive = as.character(NA), sdrf = as.character(NA),
  idf = as.character(NA), adf = as.character(NA),
  gpl_dir = as.character(NA), gpl_files = as.character(NA),
  annot_gpl = as.logical(NA))
}
\arguments{
\item{acc}{ABC}

\item{import_method}{DEF}

\item{normalise_method}{FED}

\item{raw_dir}{GHI}

\item{raw_files}{JKL}

\item{raw_archive}{MNO}

\item{processed_dir}{PQR}

\item{processed_files}{STU}

\item{processed_archive}{987}

\item{sdrf}{VWX}

\item{idf}{YZ0}

\item{adf}{123}

\item{gpl_dir}{456}

\item{gpl_files}{789}

\item{annot_gpl}{Logical.}
}
\description{
Sets up a class for storing the options for importing datasets from GEO
and/or ArrayExpress
}
