% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_classes.R
\name{MicroarrayPreprocessConfig-class}
\alias{MicroarrayPreprocessConfig-class}
\title{Class definition for MicroarrayPreprocessConfig}
\description{
Class definition for MicroarrayPreprocessConfig
}
