% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{multisymbol_to_entrez_ids}
\alias{multisymbol_to_entrez_ids}
\title{multisymbol_to_entrez_ids}
\usage{
multisymbol_to_entrez_ids(gene_symbols = character(0),
  symbol_type = "SYMBOL", entrezgene_db = NULL,
  split_character = "\\\\|", collapse_character = "|", quiet = TRUE)
}
\arguments{
\item{gene_symbols}{vector of gene_symbols of a single type.}

\item{symbol_type}{type of the gene_symbols (eg, Refseq).}

\item{entrezgene_db}{database for mapping gene_symbols to
entrez.ids.}

\item{split_character}{character string for splitting up the
symbols if any of the input entries contains more than one symbol.}

\item{collapse_character}{character for collapsing entrez ids.}

\item{quiet}{BOOLEAN should dplyr progress bar be suppressed
(sideeffect).}
}
\description{
A function to map a vector of gene symbols to Entrez gene ids.
The input may contain "multiple|joined|symbols", and if it does, the output
for that entry will contain a "joined|string|of|ids" containing all of the
unique values that any of the input symbols mapped to
See \code{symbol_to_entrez_id} for further details
}
