% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation_functions.R
\name{normalise_raw_affy_eset}
\alias{normalise_raw_affy_eset}
\title{Take newly imported affymetrix data and rma-normalise then apply some
filters to the normalised expression data}
\usage{
normalise_raw_affy_eset(exprs)
}
\arguments{
\item{exprs}{Newly imported Affymetrix data}
}
\description{
Take newly imported affymetrix data and rma-normalise then apply some
filters to the normalised expression data
}
