% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{paste_gene_symbols}
\alias{paste_gene_symbols}
\title{Function to zip together a vector of entrez_ids with the gene symbol
  corresponding to each. The returned vector is in the form
  c("10000|AKT3", "1234|CCR5", ...)}
\usage{
paste_gene_symbols(entrez_ids, entrezgene_db, collapse_character = "|")
}
\arguments{
\item{entrez_ids}{a vector of NCBI entrez gene ids.}

\item{entrezgene_db}{an AnnotationDbi-style database for use in
obtaining the symbols corresponding to the gene ids.}

\item{collapse_character}{a single character for use in collapsing
gene.id and gene.symbol into "gene.id|gene.symbol".}
}
\value{
a vector of "gene.id|gene.symbol" entries.
}
\description{
Function to zip together a vector of entrez_ids with the gene symbol
  corresponding to each. The returned vector is in the form
  c("10000|AKT3", "1234|CCR5", ...)
}
