% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_classes.R
\name{preprocess_eset_workflow}
\alias{preprocess_eset_workflow}
\title{Workflow function for converting an ExpressionSet into an entrez.id
  annotated ExpressionSet that has been log transformed / quantile
  normalised and median-normalised. Probes that do not map to entrez.id are
  dropped by default. All annotation / filtering functions can be overridden}
\usage{
preprocess_eset_workflow(gset = NULL, entrezgene.db.id = NULL,
  eset.annot.fn = gld_fnDefault_esetAnnotation,
  keep.sample.fn = gld_fnDefault_keepSample,
  transform.and.filter.fn = gld_fnDefault_transformAndProbeFilter,
  keep.probe.fn = gld_fnDefault_keepProbe)
}
\arguments{
\item{gset}{An ExpressionSet (must have genbank
ids or swissprot ids or similar for default annotation function).}

\item{entrezgene.db.id}{character - Indicates which
AnnotationDbi object to use in the esetAnnotation function.}

\item{eset.annot.fn}{Function that annotates an input
ExpressionSet and outputs an ExpressionSet.}

\item{keep.sample.fn}{Function that returns integer
indices of samples (cols) of the ExpressionSet that are to be kept.}

\item{transform.and.filter.fn}{Function that applies various
filtering / transformation steps to the values in the input ExpressionSet
and returns an ExpressionSet}

\item{keep.probe.fn}{Function that returns integer
indices of probes (rows) of the ExpressionSet that are to be kept.}
}
\value{
An ExpressionSet
}
\description{
Workflow function for converting an ExpressionSet into an entrez.id
  annotated ExpressionSet that has been log transformed / quantile
  normalised and median-normalised. Probes that do not map to entrez.id are
  dropped by default. All annotation / filtering functions can be overridden
}
