% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_classes.R
\docType{methods}
\name{run_annotate_and_preprocess_workflow}
\alias{run_annotate_and_preprocess_workflow}
\alias{run_annotate_and_preprocess_workflow,ExpressionSet,MicroarrayImportConfig,character-method}
\title{Generic for the method 'run_annotate_and_preprocess_workflow'}
\description{
Generic for the method 'run_annotate_and_preprocess_workflow'

run_annotate_and_preprocess_workflow for ExpressionSet and
  MicroarrayImportConfig
}
