% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffex_methods.R
\docType{methods}
\name{run_diffex_workflow}
\alias{run_diffex_workflow}
\alias{run_diffex_workflow,ExpressionSet,DiffexConfig-method}
\title{Generic for the method 'run_diffex_workflow'}
\value{
a list with entries (design, contrast, fits_init, fits)
}
\description{
Generic for the method 'run_diffex_workflow'

'run_diffex_workflow' for ExpressionSet and DiffexConfig
}
