% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_classes.R
\docType{methods}
\name{run_download_workflow}
\alias{run_download_workflow}
\alias{run_download_workflow,MicroarrayDownloadConfig-method}
\title{Generic for the method 'run_download_workflow'}
\description{
Generic for the method 'run_download_workflow'

'run_download_workflow' method for the MicroarrayDownloadConfig class.
}
