% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_methods.R
\name{run_download_workflow}
\alias{run_download_workflow}
\title{'run_download_workflow' method for the MicroarrayDownloadConfig class.}
\usage{
run_download_workflow(object)
}
\arguments{
\item{object}{A MicroarrayDownloadConfig object}
}
\description{
'run_download_workflow' method for the MicroarrayDownloadConfig class.
}
