% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_methods.R
\docType{methods}
\name{run_import_workflow}
\alias{run_import_workflow}
\alias{run_import_workflow,MicroarrayImportConfig-method}
\title{Generic for the method 'run_import_workflow'}
\value{
a Biobase::ExpressionSet
}
\description{
Generic for the method 'run_import_workflow'

'run_import_workflow' method for the MicroarrayImportConfig class
}
