% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_methods.R
\docType{methods}
\name{run_preprocess_workflow}
\alias{run_preprocess_workflow}
\alias{run_preprocess_workflow,ExpressionSet,MicroarrayPreprocessConfig-method}
\title{Generic for the method 'run_preprocess_workflow'}
\value{
a Biobase::ExpressionSet
}
\description{
Generic for the method 'run_preprocess_workflow'

'run_preprocess_workflow' for ExpressionSet and MicroarrayPreprocessConfig
}
