% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_functions.R
\name{swissprot_column_to_refseq}
\alias{swissprot_column_to_refseq}
\title{Extract any NM_XXXX or NR_XXXX RefSeq/GenBank identifiers out of a SwissProt
column in an \code{ExpressionSet}}
\usage{
swissprot_column_to_refseq(swiss)
}
\arguments{
\item{swiss}{A vector of swissprot ids.}
}
\description{
Cells are delimited as follows: ABC /// DEF /// GHI1 // GHI2 /// JKL
So that main entries are ///-separated and subentries are //-separated.
}
\details{
Returned value should be a vector of comma-separated refseq ids.
}
