% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{symbol_to_entrez_id}
\alias{symbol_to_entrez_id}
\title{Function to map gene symbols to entrez gene ids. The input gene symbols can
be of a variety of formats (eg, REFSEQ, ENSEMBL) provided the symbol type is
present as a column name in the provided AnnotationDbi dataset.
If any symbol maps to multiple entrez gene ids, the entrez ids are collapsed
into a single string.
The function returns a vector of entrez ids, with the order of the output
matching exactly with the order of the input.
Any symbol that can"t be mapped to an entrez id returns with the value NA}
\usage{
symbol_to_entrez_id(gene_symbols = character(0),
  symbol_type = "SYMBOL", entrezgene_db = NULL,
  collapse_character = "|")
}
\arguments{
\item{gene_symbols}{vector of some type of gene symbol (eg,
Ensembl, Refseq, HGNC)}

\item{symbol_type}{the type of gene symbol used (should be a
colname for the entrezgene_db)}

\item{entrezgene_db}{a database object for use in mapping symbol to
entrezGene id.}

\item{collapse_character}{a single character for collapsing
multiple ids into a single entry.}
}
\value{
a vector of the same length as gene_symbols containing the
  entrez ids that correspond to the input gene_symbols.
}
\description{
Function to map gene symbols to entrez gene ids. The input gene symbols can
be of a variety of formats (eg, REFSEQ, ENSEMBL) provided the symbol type is
present as a column name in the provided AnnotationDbi dataset.
If any symbol maps to multiple entrez gene ids, the entrez ids are collapsed
into a single string.
The function returns a vector of entrez ids, with the order of the output
matching exactly with the order of the input.
Any symbol that can"t be mapped to an entrez id returns with the value NA
}
