% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_classes.R
\name{gld_fnDefault_esetAnnotation}
\alias{gld_fnDefault_esetAnnotation}
\title{Default function for annotating the ExpressionSet with additional info}
\usage{
gld_fnDefault_esetAnnotation(geo.limma.dataset = NULL, gset = NULL,
  entrezgene.db.id = NULL)
}
\arguments{
\item{geo.limma.dataset}{A geo_limma_dataset.}

\item{gset}{An ExpressionSet.}

\item{entrezgene.db.id}{The id for the AnnotationDBI for use in
mapping to entrez ids.}
}
\description{
This just adds entrez.ids.
Not exported.
Not sure how to import entrezgene database (entrezgene.db.id) by name
  when using \<AT\>import annotations.
}
