% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_classes.R
\name{preprocess_eset_workflow}
\alias{preprocess_eset_workflow}
\title{Workflow function for converting an ExpressionSet into an entrez.id
  annotated ExpressionSet that has been log transformed / quantile
  normalised and median-normalised. Probes that do not map to entrez.id are
  dropped by default. All annotation / filtering functions can be overridden}
\usage{
preprocess_eset_workflow(gset = NULL, entrezgene_db = NULL,
  eset_annot_fn = gld_fnDefault_esetAnnotation,
  keep_sample_fn = gld_fnDefault_keepSample,
  transform_and_filter_fn = gld_fnDefault_transformAndProbeFilter,
  keep_probe_fn = gld_fnDefault_keepProbe)
}
\arguments{
\item{gset}{An ExpressionSet (must have genbank ids or
swissprot ids or similar for default annotation function).}

\item{entrezgene_db}{character or \code{OrgDb}. Which
\code{AnnotationDbi} object to use in the eset annotation function.}

\item{eset_annot_fn}{Function that annotates an input ExpressionSet
and outputs an ExpressionSet.}

\item{keep_sample_fn}{Function that returns integer indices of
samples (cols) of the ExpressionSet that are to be kept.}

\item{transform_and_filter_fn}{Function that applies various
filtering / transformation steps to the values in the input ExpressionSet
and returns an ExpressionSet}

\item{keep_probe_fn}{Function that returns integer indices of
probes (rows) of the ExpressionSet that are to be kept.}
}
\value{
An ExpressionSet
}
\description{
Workflow function for converting an ExpressionSet into an entrez.id
  annotated ExpressionSet that has been log transformed / quantile
  normalised and median-normalised. Probes that do not map to entrez.id are
  dropped by default. All annotation / filtering functions can be overridden
}
